% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{api.players.by.game.type}
\alias{api.players.by.game.type}
\title{api.players.by.game.type}
\usage{
api.players.by.game.type(season, team, game.type)
}
\arguments{
\item{season}{season string}

\item{team}{triCode}

\item{game.type}{choose one ("regular.season" or 2) ("playoffs" or 3)}
}
\value{
dataframe
}
\description{
api.players.by.game.type
}
\examples{
api.players.by.game.type("20242025", "STL", "regular.season")
}
