% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{api.teams.by.season}
\alias{api.teams.by.season}
\title{api.teams.by.season}
\usage{
api.teams.by.season(season)
}
\arguments{
\item{season}{season string}
}
\value{
dataframe
}
\description{
api.teams.by.season
}
\examples{
api.teams.by.season("20242025")
}
