% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{game.id.object-class}
\alias{game.id.object-class}
\title{game.id.object}
\description{
game.id.object
}
\section{Slots}{

\describe{
\item{\code{game.stats}}{ANY.}

\item{\code{player.stats.by.game}}{ANY.}

\item{\code{plays.by.game}}{ANY.}
}}

