% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{general.league.information-class}
\alias{general.league.information-class}
\title{general.league.information}
\description{
general.league.information
}
\section{Slots}{

\describe{
\item{\code{teams.in.league}}{ANY.}

\item{\code{players.in.league}}{ANY.}

\item{\code{games.played}}{ANY.}
}}

