% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{league-class}
\alias{league-class}
\title{league}
\description{
league
}
\section{Slots}{

\describe{
\item{\code{teams}}{data.frame; teams}

\item{\code{players}}{data.frame; players}

\item{\code{all.games}}{data.frame; games}
}}

