% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{preseason-class}
\alias{preseason-class}
\title{preseason class}
\description{
preseason class
}
\section{Slots}{

\describe{
\item{\code{game.stats}}{data.frame; game stats}

\item{\code{games}}{list; games}
}}

