% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/generics.methods.R
\docType{class}
\name{SLAP-class}
\alias{SLAP-class}
\alias{show,SLAP-method}
\title{SLAP}
\usage{
\S4method{show}{SLAP}(object)
}
\arguments{
\item{object}{a \code{SLAPbase} object}
}
\description{
SLAP
}
\section{Methods (by generic)}{
\itemize{
\item \code{show(SLAP)}: compact print

}}
\section{Slots}{

\describe{
\item{\code{league}}{S4; league object}

\item{\code{preseason}}{S4; preseason object}

\item{\code{regular.season}}{S4; regular.season object}

\item{\code{playoffs}}{S4; playoffs object}
}}

