% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{SLAP}
\alias{SLAP}
\title{SLAP}
\usage{
SLAP(
  season = current.season(),
  team,
  game.types = c("preseason", "regular.season", "playoffs")
)
}
\arguments{
\item{season}{season string}

\item{team}{teamName, teamId, or triCode}

\item{game.types}{list; select one or more c("preseason", "regular.season", "playoffs")}
}
\value{
SLAP object
}
\description{
SLAP
}
\examples{
SLAP("20242025", "STL")
}
