% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionality.R
\name{SLAP.season}
\alias{SLAP.season}
\title{SLAP.season}
\usage{
SLAP.season(SLAPbase)
}
\arguments{
\item{SLAPbase}{SLAPbase}
}
\value{
character
}
\description{
SLAP.season
}
\examples{
\dontrun{
 STL.20242025 <- SLAP(team="STL")
 SLAP.season(STL.20242025)
}
}
