% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionality.R
\name{SLAP.team}
\alias{SLAP.team}
\title{SLAP.team}
\usage{
SLAP.team(SLAPbase)
}
\arguments{
\item{SLAPbase}{SLAPbase}
}
\value{
list(teamName, teamId, triCode)
}
\description{
SLAP.team
}
\examples{
\dontrun{
 STL.20242025 <- SLAP(team="STL")
 SLAP.team(STL.20242025)
}
}
