% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionality.R
\name{SLAP.zip}
\alias{SLAP.zip}
\title{SLAP.zip}
\usage{
SLAP.zip(SLAPbase, filename = NA)
}
\arguments{
\item{SLAPbase}{SLAPbase}

\item{filename}{default is working dir/season.team.zip}
}
\value{
zip save
}
\description{
SLAP.zip
}
\examples{
\dontrun{
STL.20242025 <- SLAP(team="STL")
SLAP.zip(STL.20242025, "STL.20242025.zip")
}
}
