% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{teamName.teamId.triCode}
\alias{teamName.teamId.triCode}
\title{teamName.teamId.triCode}
\usage{
teamName.teamId.triCode(team, SLAPbase = NA)
}
\arguments{
\item{team}{teamName, teamId, or triCode}

\item{SLAPbase}{SLAPbase}
}
\value{
list(teamName, teamId, triCode)
}
\description{
teamName.teamId.triCode
}
\examples{
teamName.teamId.triCode("STL")
}
