% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{api.teams.by.season}
\alias{api.teams.by.season}
\title{api.teams.by.season}
\usage{
api.teams.by.season(season, all.teams = FALSE)
}
\arguments{
\item{season}{\code{string} Season identifier.}

\item{all.teams}{\code{logical(1)} Include all past and current teams? Default \code{FALSE}}
}
\value{
'data.frame' A table containing team identifiers.
}
\description{
Pulls the teams active in the provided season.
}
\examples{
api.teams.by.season("20242025")
}
