% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{playoffs-class}
\alias{playoffs-class}
\title{playoffs class}
\description{
playoffs class
}
\section{Slots}{

\describe{
\item{\code{player.stats}}{data.frame; player stats}

\item{\code{game.stats}}{data.frame; game stats}

\item{\code{games}}{list; games}
}}

