% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{regular.season-class}
\alias{regular.season-class}
\title{regular.season class}
\description{
regular.season class
}
\section{Slots}{

\describe{
\item{\code{players.by.game.type}}{ANY.}

\item{\code{games.played}}{ANY.}
}}

