% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{SLAP}
\alias{SLAP}
\title{SLAP}
\usage{
SLAP(
  season = current.season(),
  team,
  game.types = c("pre.season", "regular.season", "playoffs")
)
}
\arguments{
\item{season}{season string}

\item{team}{}
}
\value{
SLAP object
}
\description{
SLAP
}
\examples{
SLAP("20242025", "STL")
}
